chore: reorganize pages and add new news and 3d render pages
