+++ src/App.css
@@ -25,12 +25,6 @@
   z-index: 1;
 }
 
-header {
-  display: flex;
-  margin: 0;
-  flex-direction: column;
-}
-
 .swiper-wrapper {
   width: 100%;
 }
@@ -117,7 +111,6 @@ header > p {
   display: flex;
   justify-content: space-between;
   align-items: center;
-  flex-direction: column;
 }
 
 .fact,
@@ -246,9 +239,7 @@ section.visor h2 {
   font-size: 0.8rem;
 }
 .alien-popup div {
-  background-color: rgba(173, 121, 236, 0.2);
-  -webkit-backdrop-filter: blur(20px);
-  backdrop-filter: blur(20px);
+  background-color: rgba(173, 121, 236, 1);
   box-shadow: 2px 2px 20px #ad79ec;
 }
