+++ src/App.jsx
@@ -6,20 +6,23 @@ import {
   createRoutesFromElements,
 } from "react-router-dom";
 import "animate.css";
+
 import HeroPage from "./pages/Hero";
 import ExplorePage from "./pages/Explore";
-import VenusPage from "./pages/Venus";
-import MarsPage from "./pages/Mars";
-import UranusPage from "./pages/Uranus";
-import MercuryPage from "./pages/Mercury";
-import JupiterPage from "./pages/Jupiter";
-import NeptunePage from "./pages/Neptune";
-import PlutoPage from "./pages/Pluto";
-import SaturnPage from "./pages/Saturn";
-import KupierPage from "./pages/Kupier";
-import AsteroidPage from "./pages/Asteroid";
 import ReturnPage from "./pages/Return";
-import SunPage from "./pages/Sun";
+import VenusPage from "./pages/planets/Venus";
+import MarsPage from "./pages/planets/Mars";
+import UranusPage from "./pages/planets/Uranus";
+import MercuryPage from "./pages/planets/Mercury";
+import JupiterPage from "./pages/planets/Jupiter";
+import NeptunePage from "./pages/planets/Neptune";
+import PlutoPage from "./pages/planets/Pluto";
+import SaturnPage from "./pages/planets/Saturn";
+import KupierPage from "./pages/other_celestials/Kupier";
+import AsteroidPage from "./pages/other_celestials/Asteroid";
+import SunPage from "./pages/other_celestials/Sun";
+import Render from "./pages/Render";
+import News from "./pages/News";
 
 function App() {
   const router = createBrowserRouter(
@@ -52,6 +55,10 @@ function App() {
         <Route path="/return" element={<ReturnPage />} />
 
         <Route path="/sun" element={<SunPage />} />
+
+        <Route path="/render" element={<Render />} />
+
+        <Route path="/news" element={<News />} />
       </>
     )
   );
