+++ src/pages/News.jsx
@@ -1,0 +1,10 @@
+const News = () => {
+  return (
+    <div>
+      <h1>News</h1>
+      <h2>Picture of the day</h2>
+    </div>
+  );
+};
+
+export default News;
