+++ src/pages/Render.jsx
@@ -1,0 +1,9 @@
+const Render = () => {
+  return (
+    <div>
+      <h1>Render</h1>
+    </div>
+  );
+};
+
+export default Render;
