+++ src/pages/other_celestials/Asteroid.jsx
@@ -1,7 +1,7 @@
 import { Link } from "react-router-dom";
 
-import { BeltLayout } from "../components/layouts/BeltLayout";
-import { FactLayout } from "../components/layouts/FactLayout";
+import { BeltLayout } from "../../components/layouts/BeltLayout";
+import { FactLayout } from "../../components/layouts/FactLayout";
 
 import AsteroidBelt from "/assets/images/asteroid/asteroid_belt.webp";
 import Ceres from "/assets/images/asteroid/ceres.jpeg";
